python-watermark-app/requirements.txt
Pillow==10.0.0
# tkinter is included with Python standard library
# No additional GUI dependencies needed